refactor: :recycle: move tui into separate folder
